AS-3092: Implement ImportEndpointsImpl.

+++ farms-api/farms-persistence/src/main/java/ca/bc/gov/farms/persistence/v1/dao/mybatis/ImportDaoImpl.java
@@ -31,7 +31,7 @@ public class ImportDaoImpl extends BaseDao implements ImportDao {
 
         try (CallableStatement callableStatement = this.conn
                 .prepareCall("call farms_webapp_pkg.insert_import_version(?, ?, ?, ?, ?, ?, ?, ?)")) {
-            callableStatement.registerOutParameter(1, Types.NUMERIC);
+            callableStatement.registerOutParameter(1, Types.BIGINT);
             callableStatement.setString(2, dto.getImportClassCode());
             callableStatement.setString(3, dto.getImportStateCode());
             callableStatement.setString(4, dto.getDescription());
